+++ Libplanet.Explorer.Executable/Program.cs
@@ -280,7 +280,7 @@ If omitted (default) explorer only the local blockchain store.")]
         private static IRichStore LoadStore(Options options)
         {
             // FIXME: This method basically does the same thing to Libplanet.Extensions.Cocona's
-            // StoreCommand.LoadStoreFromUri() & StatsCommand.LoadStoreFromUri() methods.
+            // Utils.LoadStoreFromUri() method.
             // The duplicate code should be extract to a shared common method.
             // https://github.com/planetarium/libplanet/issues/1573
             bool readOnlyMode = options.Seeds is null;